Refactor: 컴포넌트 분리 및 타입 import
