+++ src/pages/playlistDetail/Player.tsx
@@ -1,0 +1,82 @@
+import { Playlist } from "../../types/playlist";
+import { Video } from "../../types/video";
+import PlaylistActions from "../../components/common/PlaylistAction";
+
+export const Player = ({ playlist, video }: { playlist: Playlist; video: Video }) => {
+  // youtube URL을 embed용 URL로 변환
+  const getEmbedUrl = (url: string): string => {
+    try {
+      const urlObj = new URL(url);
+      const hostname = urlObj.hostname;
+      let videoId = "";
+
+      if (hostname === "youtu.be") {
+        videoId = urlObj.pathname.slice(1);
+      } else if (hostname === "www.youtube.com" || hostname === "youtube.com") {
+        const queryVideoId = urlObj.searchParams.get("v");
+        if (queryVideoId) {
+          videoId = queryVideoId;
+        }
+      }
+
+      if (!videoId) {
+        return url; // 예외 처리
+      }
+
+      const params = new URLSearchParams({
+        autoplay: "1", // 자동재생
+        mute: "1", // 음소거(자동재생 하려면 필수)
+        controls: "1", // 컨트롤러 숨기려면 0
+        modestbranding: "1", // 유튜브 로고 최소화
+        rel: "0", // 관련 영상 안 보이도록
+      });
+
+      return `https://www.youtube.com/embed/${videoId}?${params.toString()}`;
+    } catch {
+      console.warn("잘못된 URL 형식입니다:", url);
+      return "";
+    }
+  };
+
+  const embedUrl = getEmbedUrl(video.url);
+
+  return (
+    <>
+      {/* 영상 영역 */}
+      <div className="relative w-full pt-[56.25%]">
+        {/* 16:9 비율 */}
+        <iframe
+          className="absolute left-0 top-0 h-full w-full"
+          src={embedUrl}
+          title={video.title}
+          allowFullScreen
+        />
+      </div>
+
+      {/* 콘텐츠 영역 */}
+      <div className="px-4 pb-6 pt-3">
+        {/* 유저 정보 */}
+        <div className="flex flex-row items-center justify-between">
+          <div className="flex flex-row gap-2.5">
+            <img
+              src="https://i.pinimg.com/736x/17/c1/d9/17c1d903910937ecfd18943ee06279c2.jpg"
+              alt="ijisun 프로필 이미지"
+              className="h-6 w-6 rounded-full"
+            />
+            <p>ijisun</p>
+          </div>
+          <div className="text-sub">
+            <p>등록일 {playlist.created_at}</p>
+          </div>
+        </div>
+
+        {/* 플레이리스트 정보 */}
+        <div className="pt-4">
+          <h3 className="text-body1-bold">{playlist.title}</h3>
+          <p className="mb-4 mt-2 text-sub2">{playlist.description}</p>
+          <PlaylistActions />
+        </div>
+      </div>
+    </>
+  );
+};
